+++ ros_gz_bridge/src/factory_interface.hpp
@@ -24,6 +24,8 @@
 // include ROS 2
 #include <rclcpp/rclcpp.hpp>
 
+#include "bridge_handle_gz_to_ros_parameters.hpp"
+
 namespace ros_gz_bridge
 {
 
@@ -37,7 +39,7 @@ public:
   create_ros_publisher(
     rclcpp::Node::SharedPtr ros_node,
     const std::string & topic_name,
-    size_t queue_size) = 0;
+    const rclcpp::QoS & qos) = 0;
 
   virtual
   gz::transport::Node::Publisher
@@ -51,7 +53,7 @@ public:
   create_ros_subscriber(
     rclcpp::Node::SharedPtr ros_node,
     const std::string & topic_name,
-    size_t queue_size,
+    const rclcpp::QoS & qos,
     gz::transport::Node::Publisher & gz_pub) = 0;
 
   virtual
@@ -61,7 +63,7 @@ public:
     const std::string & topic_name,
     size_t queue_size,
     rclcpp::PublisherBase::SharedPtr ros_pub,
-    bool override_timestamps_with_wall_time) = 0;
+    const BridgeHandleGzToRosParameters & gz_to_ros_parameters) = 0;
 };
 
 }  // namespace ros_gz_bridge