+++ ros_gz_bridge/test/utils/ros_test_msg.cpp
@@ -20,6 +20,8 @@
 #include <string>
 #include <cstddef>
 
+#include <marine_acoustic_msgs/msg/dvl.hpp>
+
 #include "gz/msgs/config.hh"
 
 #if GZ_MSGS_MAJOR_VERSION >= 10
@@ -562,6 +564,77 @@ void compareTestMsg(const std::shared_ptr<gps_msgs::msg::GPSFix> & _msg)
 
   for (auto i = 0u; i < 9; ++i) {
     EXPECT_FLOAT_EQ(0, _msg->position_covariance[i]);
+  }
+}
+
+void createTestMsg(marine_acoustic_msgs::msg::Dvl & _msg)
+{
+  createTestMsg(_msg.header);
+  _msg.velocity_mode = marine_acoustic_msgs::msg::Dvl::DVL_MODE_BOTTOM;
+  _msg.dvl_type = marine_acoustic_msgs::msg::Dvl::DVL_TYPE_PISTON;
+
+  createTestMsg(_msg.velocity);
+  for (auto i = 0; i < 9; ++i) {
+    _msg.velocity_covar[i] = i;
+  }
+
+  _msg.altitude = 10;
+  _msg.course_gnd = std::atan2(_msg.velocity.y, _msg.velocity.x);
+  _msg.speed_gnd = std::sqrt(_msg.velocity.x * _msg.velocity.x + _msg.velocity.y *
+      _msg.velocity.y);
+
+  _msg.num_good_beams = 4;
+  _msg.sound_speed = -1;
+  _msg.beam_ranges_valid = true;
+  _msg.beam_velocities_valid = true;
+
+  double vLen = std::sqrt(1.0 * 1.0 + 2.0 * 2.0 + 3.0 * 3.0);
+  for (auto i = 0; i < _msg.num_good_beams; ++i) {
+    // beam_unit_vec = normalized velocity direction (1,2,3)/sqrt(14)
+    _msg.beam_unit_vec[i].x = 1.0 / vLen;
+    _msg.beam_unit_vec[i].y = 2.0 / vLen;
+    _msg.beam_unit_vec[i].z = 3.0 / vLen;
+    _msg.range[i] = 10;
+    _msg.range_covar[i] = 11;
+    _msg.beam_quality[i] = 100;
+    _msg.beam_velocity[i] = vLen;
+    _msg.beam_velocity_covar[i] = -1;
+  }
+}
+
+void compareTestMsg(const std::shared_ptr<marine_acoustic_msgs::msg::Dvl> & _msg)
+{
+  marine_acoustic_msgs::msg::Dvl expected_msg;
+  createTestMsg(expected_msg);
+
+  compareTestMsg(_msg->header);
+  EXPECT_EQ(expected_msg.velocity_mode, _msg->velocity_mode);
+  EXPECT_EQ(expected_msg.dvl_type, _msg->dvl_type);
+
+  compareTestMsg(std::make_shared<geometry_msgs::msg::Vector3>(_msg->velocity));
+  for (auto i = 0; i < 9; ++i) {
+    EXPECT_DOUBLE_EQ(expected_msg.velocity_covar[i], _msg->velocity_covar[i]);
+  }
+
+  EXPECT_DOUBLE_EQ(expected_msg.altitude, _msg->altitude);
+  EXPECT_DOUBLE_EQ(expected_msg.course_gnd, _msg->course_gnd);
+  EXPECT_DOUBLE_EQ(expected_msg.speed_gnd, _msg->speed_gnd);
+
+  ASSERT_EQ(expected_msg.num_good_beams, _msg->num_good_beams);
+  ASSERT_DOUBLE_EQ(expected_msg.sound_speed, _msg->sound_speed);
+  EXPECT_EQ(expected_msg.beam_ranges_valid, _msg->beam_ranges_valid);
+  EXPECT_EQ(expected_msg.beam_velocities_valid, _msg->beam_velocities_valid);
+
+  for (auto i = 0; i < _msg->num_good_beams; ++i) {
+    // Use EXPECT_NEAR for beam_unit_vec due to floating-point normalization.
+    EXPECT_NEAR(expected_msg.beam_unit_vec[i].x, _msg->beam_unit_vec[i].x, 1e-6);
+    EXPECT_NEAR(expected_msg.beam_unit_vec[i].y, _msg->beam_unit_vec[i].y, 1e-6);
+    EXPECT_NEAR(expected_msg.beam_unit_vec[i].z, _msg->beam_unit_vec[i].z, 1e-6);
+    EXPECT_DOUBLE_EQ(expected_msg.range[i], _msg->range[i]);
+    EXPECT_DOUBLE_EQ(expected_msg.range_covar[i], _msg->range_covar[i]);
+    EXPECT_DOUBLE_EQ(expected_msg.beam_quality[i], _msg->beam_quality[i]);
+    EXPECT_NEAR(expected_msg.beam_velocity[i], _msg->beam_velocity[i], 1e-6);
+    EXPECT_DOUBLE_EQ(expected_msg.beam_velocity_covar[i], _msg->beam_velocity_covar[i]);
   }
 }
 
@@ -1675,6 +1748,95 @@ void compareTestMsg(const std::shared_ptr<ros_gz_interfaces::msg::LogicalCameraI
     EXPECT_EQ(expected_msg.model[i].name, _msg->model[i].name);
     compareTestMsg(std::make_shared<geometry_msgs::msg::Pose>(_msg->model[i].pose));
   }
+}
+
+void createTestMsg(ros_gz_interfaces::msg::LogPlaybackStatistics & _msg)
+{
+  std_msgs::msg::Header header_msg;
+  createTestMsg(header_msg);
+  _msg.header = header_msg;
+
+  builtin_interfaces::msg::Time start_time;
+  createTestMsg(start_time);
+  _msg.start_time = start_time;
+
+  builtin_interfaces::msg::Time end_time;
+  createTestMsg(end_time);
+  _msg.end_time = end_time;
+}
+
+void compareTestMsg(const std::shared_ptr<ros_gz_interfaces::msg::LogPlaybackStatistics> & _msg)
+{
+  ros_gz_interfaces::msg::LogPlaybackStatistics expected_msg;
+  createTestMsg(expected_msg);
+
+  compareTestMsg(std::make_shared<std_msgs::msg::Header>(_msg->header));
+  compareTestMsg(std::make_shared<builtin_interfaces::msg::Time>(_msg->start_time));
+  compareTestMsg(std::make_shared<builtin_interfaces::msg::Time>(_msg->end_time));
+}
+
+void createTestMsg(ros_gz_interfaces::msg::WorldStatistics & _msg)
+{
+  std_msgs::msg::Header header_msg;
+  createTestMsg(header_msg);
+  _msg.header = header_msg;
+
+  builtin_interfaces::msg::Time sim_time;
+  createTestMsg(sim_time);
+  _msg.sim_time = sim_time;
+
+  builtin_interfaces::msg::Time pause_time;
+  createTestMsg(pause_time);
+  _msg.pause_time = pause_time;
+
+  builtin_interfaces::msg::Time real_time;
+  createTestMsg(real_time);
+  _msg.real_time = real_time;
+
+  _msg.paused = false;
+
+  _msg.iterations = 123;
+
+  _msg.model_count = 10;
+
+  ros_gz_interfaces::msg::LogPlaybackStatistics log_playback_statistics;
+  createTestMsg(log_playback_statistics);
+  _msg.log_playback_statistics = log_playback_statistics;
+
+  _msg.real_time_factor = 0.75;
+
+  builtin_interfaces::msg::Time step_size;
+  createTestMsg(step_size);
+  _msg.step_size = step_size;
+
+  _msg.stepping = true;
+}
+
+void compareTestMsg(const std::shared_ptr<ros_gz_interfaces::msg::WorldStatistics> & _msg)
+{
+  ros_gz_interfaces::msg::WorldStatistics expected_msg;
+  createTestMsg(expected_msg);
+
+  compareTestMsg(std::make_shared<std_msgs::msg::Header>(_msg->header));
+
+  compareTestMsg(std::make_shared<builtin_interfaces::msg::Time>(_msg->sim_time));
+  compareTestMsg(std::make_shared<builtin_interfaces::msg::Time>(_msg->pause_time));
+  compareTestMsg(std::make_shared<builtin_interfaces::msg::Time>(_msg->real_time));
+
+  EXPECT_EQ(expected_msg.paused, _msg->paused);
+
+  EXPECT_EQ(expected_msg.iterations, _msg->iterations);
+
+  EXPECT_EQ(expected_msg.model_count, _msg->model_count);
+
+  compareTestMsg(std::make_shared<ros_gz_interfaces::msg::LogPlaybackStatistics>(
+        _msg->log_playback_statistics));
+
+  EXPECT_FLOAT_EQ(expected_msg.real_time_factor, _msg->real_time_factor);
+
+  compareTestMsg(std::make_shared<builtin_interfaces::msg::Time>(_msg->step_size));
+
+  EXPECT_EQ(expected_msg.stepping, _msg->stepping);
 }
 
 }  // namespace testing