+++ ros_gz_sim/launch/gz_server.launch.py
@@ -40,6 +40,12 @@ def generate_launch_description():
     declare_start_paused_cmd = DeclareLaunchArgument(
         'start_paused', default_value='False',
         description='Start simulation paused if True')
+    declare_initial_sim_time_cmd = DeclareLaunchArgument(
+        'initial_sim_time', default_value='0.0',
+        description='The initial simulation time')
+    declare_verbosity_level_cmd = DeclareLaunchArgument(
+        'verbosity_level', default_value='4',
+        description='The verbosity level of the Gazebo server (0=FATAL, 4=DEBUG)')
 
     gz_server_action = GzServer(
         world_sdf_file=LaunchConfiguration('world_sdf_file'),
@@ -48,6 +54,8 @@ def generate_launch_description():
         create_own_container=LaunchConfiguration('create_own_container'),
         use_composition=LaunchConfiguration('use_composition'),
         start_paused=LaunchConfiguration('start_paused'),
+        initial_sim_time=LaunchConfiguration('initial_sim_time'),
+        verbosity_level=LaunchConfiguration('verbosity_level'),
     )
 
     # Create the launch description and populate
@@ -60,6 +68,8 @@ def generate_launch_description():
     ld.add_action(declare_create_own_container_cmd)
     ld.add_action(declare_use_composition_cmd)
     ld.add_action(declare_start_paused_cmd)
+    ld.add_action(declare_initial_sim_time_cmd)
+    ld.add_action(declare_verbosity_level_cmd)
     # Add the gz_server action
     ld.add_action(gz_server_action)