+++ ros_gz_sim/ros_gz_sim/actions/gzserver.py
@@ -97,6 +97,7 @@ class GzServer(Action):
         use_composition: Union[bool, SomeSubstitutionsType] = False,
         initial_sim_time: Union[float, SomeSubstitutionsType] = 0.0,
         start_paused: Union[bool, SomeSubstitutionsType] = False,
+        verbosity_level: Union[int, SomeSubstitutionsType] = 4,
         **kwargs
     ) -> None:
         """
@@ -111,6 +112,8 @@ class GzServer(Action):
         :param: create_own_container Whether to start a ROS container when using composition.
         :param: use_composition Use composed bringup if True.
         :param: start_paused Start simulation in a paused state if True.
+        :param: initial_sim_time: The initial simulation time.
+        :param: verbosity_level: The verbosity level of the Gazebo server (0=FATAL, 4=DEBUG).
         """
         super().__init__(**kwargs)
         self.__world_sdf_file = world_sdf_file
@@ -148,6 +151,13 @@ class GzServer(Action):
             )
         else:
             self.__start_paused = normalize_typed_substitution(start_paused, bool)
+        
+        if isinstance(verbosity_level, str):
+            self.__verbosity_level = normalize_typed_substitution(
+                TextSubstitution(text=verbosity_level), int
+            )
+        else:
+            self.__verbosity_level = normalize_typed_substitution(verbosity_level, int)
 
     @classmethod
     def parse(cls, entity: Entity, parser: Parser):
@@ -181,6 +191,10 @@ class GzServer(Action):
             'start_paused', data_type=str,
             optional=True)
 
+        verbosity_level = entity.get_attr(
+            'verbosity_level', data_type=str,
+            optional=True)
+
         if isinstance(world_sdf_file, str):
             world_sdf_file = parser.parse_substitution(world_sdf_file)
             kwargs['world_sdf_file'] = world_sdf_file
@@ -209,6 +223,10 @@ class GzServer(Action):
         if isinstance(start_paused, str):
             start_paused = parser.parse_substitution(start_paused)
             kwargs['start_paused'] = start_paused
+        
+        if isinstance(verbosity_level, str):
+            verbosity_level = parser.parse_substitution(verbosity_level)
+            kwargs['verbosity_level'] = verbosity_level
 
         return cls, kwargs
 
@@ -246,7 +264,8 @@ class GzServer(Action):
                 parameters=[{'world_sdf_file': self.__world_sdf_file,
                              'world_sdf_string': self.__world_sdf_string,
                              'initial_sim_time': self.__initial_sim_time,
-                             'start_paused': self.__start_paused}],
+                             'start_paused': self.__start_paused,
+                             'verbosity_level': self.__verbosity_level}],
                 ))
 
         # Composable node with container configuration
@@ -264,7 +283,8 @@ class GzServer(Action):
                         parameters=[{'world_sdf_file': self.__world_sdf_file,
                                      'world_sdf_string': self.__world_sdf_string,
                                      'initial_sim_time': self.__initial_sim_time,
-                                     'start_paused': self.__start_paused}],
+                                     'start_paused': self.__start_paused,
+                                     'verbosity_level': self.__verbosity_level}],
                         extra_arguments=[{'use_intra_process_comms': True}],
                         ),
                     ],
@@ -283,7 +303,8 @@ class GzServer(Action):
                         parameters=[{'world_sdf_file': self.__world_sdf_file,
                                      'world_sdf_string': self.__world_sdf_string,
                                      'initial_sim_time': self.__initial_sim_time,
-                                     'start_paused': self.__start_paused}],
+                                     'start_paused': self.__start_paused,
+                                     'verbosity_level': self.__verbosity_level}],
                         extra_arguments=[{'use_intra_process_comms': True}],
                         ),
                     ],