+++ ros_gz_sim/src/gzserver.cpp
@@ -62,8 +62,9 @@ void GzServer::OnStart()
   auto world_sdf_string = this->declare_parameter("world_sdf_string", "");
   auto initial_sim_time = this->declare_parameter("initial_sim_time", 0.0);
   auto start_paused = this->declare_parameter("start_paused", false);
+  const auto verbosity_level = this->declare_parameter("verbosity_level", 4);
 
-  gz::common::Console::SetVerbosity(4);
+  gz::common::Console::SetVerbosity(verbosity_level);
   gz::sim::ServerConfig server_config;
 
   if (!world_sdf_file.empty()) {