+++ ros_gz_sim/CHANGELOG.rst
@@ -2,6 +2,23 @@
 Changelog for package ros_gz_sim
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.22 (2026-03-26)
+-------------------
+
+1.0.21 (2026-02-24)
+-------------------
+* Added descriptions to gz_server and ros_gz_sim launch files (`#838 <https://github.com/gazebosim/ros_gz/issues/838>`_) (`#840 <https://github.com/gazebosim/ros_gz/issues/840>`_)
+* Added verbosity_level parameter. (`#833 <https://github.com/gazebosim/ros_gz/issues/833>`_) (`#836 <https://github.com/gazebosim/ros_gz/issues/836>`_)
+* Contributors: mergify[bot]
+
+1.0.20 (2026-02-04)
+-------------------
+
+1.0.19 (2026-01-20)
+-------------------
+* Add dependency to ros2pkg (`#816 <https://github.com/gazebosim/ros_gz/issues/816>`_) (`#818 <https://github.com/gazebosim/ros_gz/issues/818>`_)
+* Contributors: mergify[bot]
+
 1.0.18 (2025-11-21)
 -------------------
 * Implement ROS standard simulation interfaces (backport `#790 <https://github.com/gazebosim/ros_gz/issues/790>`_) (`#802 <https://github.com/gazebosim/ros_gz/issues/802>`_)